feat: 홈 css 및 기능 변경
