+++ index.html
@@ -2,7 +2,7 @@
 <html lang="en">
   <head>
     <meta charset="UTF-8" />
-    <link rel="icon" type="image-png" href="./src/assets/icon.png" />
+    <link rel="icon" type="image-png" href="./src/assets/Vis_fvicon.png" />
     <meta name="viewport" content="width=device-width, initial-scale=1.0" />
     <title>Vis Paperlist</title>
   </head>
