+++ src/App.jsx
@@ -1,6 +1,6 @@
 import "./App.css";
 import { Route, Routes } from "react-router-dom";
-import Home from "./pages/Home";
+import Home from "./pages/Home/Home";
 import SearchResult from "./pages/SearchResult";
 import axios from "axios";
