+++ src/pages/Home/Home.css
@@ -1,0 +1,76 @@
+.search-app {
+  font-family: Arial, sans-serif;
+  height: 100%;
+  min-height: 100%;
+  min-width: 100%;
+  width: 100%;
+  padding: 2em;
+  display: flex;
+  flex-direction: column;
+  align-items: center;
+}
+
+.input-text {
+  display: flex;
+  width: 500px;
+  height: 200px;
+  flex-direction: column;
+  justify-content: center;
+  flex-shrink: 0;
+  font-size: 35px;
+  font-style: normal;
+  font-weight: 800;
+  margin: 0 0 56px;
+}
+
+.Home-search-container {
+  display: flex;
+  flex-direction: column;
+  align-items: center;
+  margin-top: 50px;
+}
+
+.paperContent-container {
+  margin-top: 20px;
+  background-color: #fff;
+  padding: 2em;
+  border-radius: 4px;
+  box-shadow: 0 0 5px rgba(0, 0, 0, 0.1);
+}
+
+.title-container {
+  width: 993px;
+  height: 121px;
+  display: flex;
+  flex-direction: row;
+  align-items: center;
+  justify-content: center;
+  margin-top: 10%;
+}
+
+.title-text {
+  color: #373a3f;
+  font-family: Inter;
+  font-size: 90px;
+  font-style: normal;
+  font-weight: 800;
+  line-height: 26px;
+}
+
+.logo {
+  width: 91px;
+  height: 25px;
+}
+
+.home-footer-container {
+  position: fixed;
+  bottom: 0;
+  width: 100%;
+  display: flex;
+  flex-direction: column;
+  align-items: center;
+}
+
+.home-footer-text {
+  color: #656565;
+}
