+++ src/pages/Home/InputContainer.css
@@ -1,0 +1,50 @@
+.input-container {
+  display: flex;
+  align-items: center;
+  justify-content: space-between;
+  width: 100%;
+  border: none;
+}
+
+.search-container {
+  display: flex;
+  align-items: center;
+  background-color: #fff;
+  border: 2px solid #ccc;
+  border-radius: 10px;
+  padding: 0.6em;
+  width: 780px;
+  height: 50px;
+  box-sizing: border-box;
+}
+
+.search-container:hover {
+  border-color: black;
+}
+
+.searchIcon-container {
+  display: flex;
+  align-items: center;
+  margin-right: 0.5em;
+}
+
+.inputName {
+  border: none;
+  outline: none;
+  flex-grow: 1;
+}
+
+.search-type {
+  margin-left: 1em;
+  background-color: #fff;
+  border: 1px solid #ccc;
+  border-radius: 6px;
+  padding: 0.5em 1em;
+  box-sizing: border-box;
+}
+
+.searchIcon {
+  width: 24.004px;
+  height: 25px;
+  margin-right: 10px;
+}
