--- src/styles/Home.css
@@ -1,80 +1,0 @@
-.search-app {
-  font-family: Arial, sans-serif;
-  background-color: #f5f5f5; /*white smoke*/
-  min-height: 100vh; /*요소의 최소 높이 지원*/
-  padding: 2em;
-}
-
-.input-text {
-  display: flex;
-  width: 500px;
-  height: 200px;
-  flex-direction: column;
-  justify-content: center;
-  flex-shrink: 0;
-  font-size: 35px;
-  font-style: normal;
-  font-weight: 800;
-}
-
-.center-fixed-container {
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-  margin-bottom: 2em;
-}
-
-.paperContent-container {
-  margin-top: 20px;
-  background-color: #fff;
-  padding: 2em;
-  border-radius: 4px;
-  box-shadow: 0 0 5px rgba(0, 0, 0, 0.1);
-}
-
-.paperContent-containerh3 {
-  margin-bottom: 1em;
-}
-
-.paperContent-container ul {
-  list-style-type: none;
-  padding-left: 0;
-}
-
-.paperContent-container li {
-  display: flex;
-  flex-direction: row;
-  justify-content: space-between;
-  align-items: center;
-  border-bottom: 1px solid #e0e0e0;
-  padding: 0.5em 0;
-}
-
-.query-item {
-  cursor: pointer;
-  color: #007bff;
-  transition: color 0.2s ease-in-out;
-}
-
-.query-item:hover {
-  color: #0056b3;
-}
-
-.date-item {
-  margin-left: 1em;
-  font-size: 0.9em;
-  color: #888;
-}
-
-.footer-container {
-  height: 5vh;
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-  margin: 0;
-}
-
-.footer-text {
-  font-size: 13px;
-  /* color: rgb(135, 135, 135); */
-}
