--- src/styles/basicComponent.css
@@ -1,15 +1,0 @@
-.header {
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-  margin-bottom: 2em;
-}
-
-.logo-container {
-  margin-bottom: 0.5em;
-}
-
-.logo-under-text {
-  font-size: 15px;
-  color: rgb(135, 135, 135);
-}
